Move email confirmation to EmailConfirmationsController
**Why**: This is clearer
