+++ app/views/devise/mailer/confirmation_instructions.html.slim
@@ -11,12 +11,12 @@ table.button.expanded.large.radius
               td
                 center
                   = link_to t('mailer.confirmation_instructions.link_text'), \
-                    sign_up_new_password_url(confirmation_token: @token), target: '_blank', \
+                    sign_up_create_email_confirmation_url(confirmation_token: @token), target: '_blank', \
                     class: 'float-center', align: 'center'
       td.expander
 
-p = link_to sign_up_new_password_url(confirmation_token: @token), \
-    sign_up_new_password_url(confirmation_token: @token), target: '_blank'
+p = link_to sign_up_create_email_confirmation_url(confirmation_token: @token), \
+    sign_up_create_email_confirmation_url(confirmation_token: @token), target: '_blank'
 
 table.spacer
   tbody
