+++ spec/features/accessibility/user_pages_spec.rb
@@ -17,16 +17,16 @@ feature 'Accessibility on pages that require authentication', :js do
       open_email(email)
       visit_in_email(t('mailer.confirmation_instructions.link_text'))
 
-      expect(current_path).to eq(sign_up_new_password_path)
+      expect(current_path).to eq(sign_up_create_email_confirmation_path)
       expect(page).to be_accessible
     end
 
     scenario 'invalid confirmation token' do
       email = 'test@example.com'
       sign_up_with(email)
-      visit sign_up_new_password_path(confirmation_token: '123456')
+      visit sign_up_create_email_confirmation_path(confirmation_token: '123456')
 
-      expect(current_path).to eq(sign_up_new_password_path)
+      expect(current_path).to eq(sign_up_create_email_confirmation_path)
       expect(page).to be_accessible
     end
   end
