+++ spec/support/features/session_helper.rb
@@ -79,7 +79,9 @@ module Features
       User.last.update(
         confirmation_token: @raw_confirmation_token, confirmation_sent_at: Time.current
       )
-      visit sign_up_new_password_path(confirmation_token: @raw_confirmation_token)
+      visit sign_up_create_email_confirmation_path(
+        confirmation_token: @raw_confirmation_token
+      )
     end
 
     def sign_up_and_2fa
